Adjusted the if statement of "analyze triangle" for a more accurate analysis of whether it is a triangle or not. The unit type of triangle's side is changed from "int" to "double". Adjusted equations, and outputs of "anglesOfTriangle".

+++ PolygonChecker/main.c
@@ -19,8 +19,8 @@ int main() {
 		{
 		case 1:
 			printf_s("Triangle selected.\n");
-			int triangleSides[3] = { 0, 0, 0 };
-			int* triangleSidesPtr = getTriangleSides(triangleSides);
+			double triangleSides[3] = { 0, 0, 0 };
+			double* triangleSidesPtr = getTriangleSides(triangleSides);
 			//printf_s("! %d\n", triangleSidesPtr[0]);
 			char* resulttriangle = analyzeTriangle(triangleSidesPtr[0], triangleSidesPtr[1], triangleSidesPtr[2]);
 			printf_s("%s\n", resulttriangle);
